requirement.txt
certifi==2018.1.18
chardet==3.0.4
cheroot==5.8.3
CherryPy==11.0.0
click==6.7
dj-database-url==0.5.0
Django==1.11.6
django-cors-headers==2.1.0
django-extensions==1.9.6
django-filter==1.0.4
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
freeze==1.0.10
idna==2.6
jdcal==1.3
olefile==0.44
openpyxl==2.4.9
Pillow==4.3.0
portend==2.2
psycopg2==2.7.3.1
PyJWT==1.6.0
pypiwin32==223
python-dotenv==0.9.1
pytz==2017.2
pywin32==223
requests==2.18.4
six==1.11.0
tempora==1.9
urllib3==1.22
